+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [0.4.13] - 2019-12-15
+### Added
+- Allow builders to change default Piping Server URLs at build-time
+
+### Changed
+- Use "./" as publicPath
+- Remove public/_redirects
+
 ## [0.4.12] - 2019-12-15
 ### Changed
 - Update dependencies
@@ -115,7 +123,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 ### Added
 - First release
 
-[Unreleased]: https://github.com/nwtgck/piping-ui-web/compare/v0.4.12...HEAD
+[Unreleased]: https://github.com/nwtgck/piping-ui-web/compare/v0.4.13...HEAD
+[0.4.13]: https://github.com/nwtgck/piping-ui-web/compare/v0.4.12...v0.4.13
 [0.4.12]: https://github.com/nwtgck/piping-ui-web/compare/v0.4.11...v0.4.12
 [0.4.11]: https://github.com/nwtgck/piping-ui-web/compare/v0.4.10...v0.4.11
 [0.4.10]: https://github.com/nwtgck/piping-ui-web/compare/v0.4.9...v0.4.10